+++ py_backend/users/views/__init__.py
@@ -7,4 +7,5 @@ from .update_profile_password import *
 from .update_profile_picture import *
 from .get_csrf_token import *
 from .get_user_datas import *
-from .username_is_unique import *
+from .unique_username import *
+from .unique_email import *
