+++ py_backend/users/views/unique_email.py
@@ -1,0 +1,22 @@
+from django.views.decorators.http import require_http_methods
+from django.views.decorators.csrf import requires_csrf_token
+from django.http import JsonResponse
+
+from users.utils import email_is_unique, decode_json_body
+
+
+@require_http_methods(["POST"])
+@requires_csrf_token
+def unique_email(request):
+    data = decode_json_body(request)
+    if isinstance(data, JsonResponse):
+        return data
+    
+    email = data.get('email')
+
+    if email:
+        is_unique, response = email_is_unique(email)
+        if is_unique:
+            return JsonResponse({'status': 'success'}, status=200)
+        return JsonResponse({'status': 'failure', 'error': response}, status=400)
+    return JsonResponse({'status': "Missing email."}, status=400)
